Basic Initial Layout
Added a message telling user to input and a box to input text in
